restore databse

+++ internal/domain/checking.go
@@ -1,7 +1,7 @@
 package domain
 
 type Checking struct {
-	AccountID     int
+	AccountID     int64
 	Account       Account
 	ServiceCharge float64
 	Amount        float64
