+++ internal/domain/saving.go
@@ -1,7 +1,7 @@
 package domain
 
 type Saving struct {
-	AccountID    int
+	AccountID    int64
 	Account      Account
 	InterestRate float64
 	Amount       int64
